Split the crates
