--- helpers/Cargo.toml
@@ -1,29 +1,0 @@
-[package]
-name = "litesvm-helpers"
-version = "0.1.0"
-edition = "2021"
-
-[features]
-default = ["loader", "token"]
-loader = []
-token = ["spl-token"]
-token-2022 = ["spl-token-2022"]
-
-[dependencies]
-spl-associated-token-account = { workspace = true, features = [
-    "no-entrypoint",
-] }
-litesvm.workspace = true
-solana-sdk.workspace = true
-spl-token-2022 = { workspace = true, optional = true, features = [
-    "no-entrypoint",
-] }
-spl-token = { workspace = true, optional = true, features = ["no-entrypoint"] }
-smallvec.workspace = true
-
-[dev-dependencies]
-test-log.workspace = true
-
-[[test]]
-name = "spl"
-required-features = ["token-2022"]
