+++ svm/Cargo.toml
@@ -1,12 +1,15 @@
 [package]
 name = "litesvm"
 description = "A fast and lightweight Solana VM simulator for testing solana programs"
-license = "Apache-2.0"
-version = "0.1.0"
-edition = "2021"
-rust-version = "1.75.0"
-repository = "https://github.com/LiteSVM/litesvm"
-readme = "README.md"
+license.workspace = true
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+repository.workspace = true
+readme = "../README.md"
+
+[features]
+internal-test = []
 
 [dependencies]
 itertools.workspace = true
@@ -36,8 +39,8 @@ test-log.workspace = true
 solana-config-program.workspace = true
 serde.workspace = true
 
-[features]
-internal-test = []
+[lints]
+workspace = true
 
 [[bench]]
 name = "banks_client_comparison"
