+++ token/src/burn.rs
