+++ token/src/burn_checked.rs
